# Feature/lazy update of time indicator on plots