--- a/package.json
+++ b/package.json
@@ -102,6 +102,7 @@
     "react-plotly.js": "^2.5.0",
     "react-redux": "^7.1.3",
     "react-router-dom": "^5.2.0",
+    "react-waypoint": "^9.0.3",
     "redux": "4.0.1",
     "redux-logic": "^2.1.1",
     "reselect": "4.0.0",