--- a/src/containers/Plot/index.tsx
+++ b/src/containers/Plot/index.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
 import { connect } from "react-redux";
+import { Waypoint } from "react-waypoint";
 import PlotlyPlot from "react-plotly.js";
 // TODO: Use changeTime action to allow updating of app time by dragging the
 // time indicator bar
@@ -18,10 +19,20 @@ interface PlotProps {
     // changeTime: ActionCreator<ChangeTimeAction>;
 }
 
+interface PlotState {
+    isInView: boolean;
+}
+
 const styles = require("./style.css");
 const TIME_INDICATOR_LINE = "timeIndicatorLine";
 
-class Plot extends React.Component<PlotProps, {}> {
+class Plot extends React.Component<PlotProps, PlotState> {
+    constructor(props: PlotProps) {
+        super(props);
+        this.state = {
+            isInView: false,
+        };
+    }
     componentDidMount() {
         const { plotConfig, time } = this.props;
         if (plotConfig.shouldRenderTimeIndicator) {
@@ -46,13 +57,23 @@ class Plot extends React.Component<PlotProps, {}> {
         }
     }
 
-    shouldComponentUpdate() {
-        return this.props.plotConfig.shouldRenderTimeIndicator;
+    shouldComponentUpdate(nextProps: PlotProps, nextState: PlotState) {
+        const { shouldRenderTimeIndicator } = this.props.plotConfig;
+        if (!shouldRenderTimeIndicator) return false;
+
+        if (this.state.isInView) {
+            return true;
+        } else if (!this.state.isInView && nextState.isInView) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public render(): JSX.Element | null {
         const { plotConfig, time } = this.props;
         const { shouldRenderTimeIndicator, data, layout } = plotConfig;
+        console.log("rendering Plot:", layout.title);
 
         if (shouldRenderTimeIndicator) {
             // Position the time indicator line at current time (the time indicator should be the last
@@ -64,19 +85,25 @@ class Plot extends React.Component<PlotProps, {}> {
         }
 
         return (
-            <div className={styles.container}>
-                <PlotlyPlot
-                    data={data}
-                    useResizeHandler={true}
-                    layout={layout}
-                    // config attributes:
-                    // https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js#L23
-                    config={{
-                        modeBarButtons: [["resetViews"]],
-                        displaylogo: false,
-                    }}
-                />
-            </div>
+            <Waypoint
+                // debug={true}
+                onEnter={() => this.setState({ isInView: true })}
+                onLeave={() => this.setState({ isInView: false })}
+            >
+                <div className={styles.container}>
+                    <PlotlyPlot
+                        data={data}
+                        useResizeHandler={true}
+                        layout={layout}
+                        // config attributes:
+                        // https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js#L23
+                        config={{
+                            modeBarButtons: [["resetViews"]],
+                            displaylogo: false,
+                        }}
+                    />
+                </div>
+            </Waypoint>
         );
     }
 }